# tests/fixtures/contracts/openapi.json
{
  "components": {
    "schemas": {
      "Engagement": {
        "additionalProperties": false,
        "description": "The top-level container that groups an authorized assessment's work.\n\nAn engagement owns its scope and authorization reference; assets, scans,\njobs, findings, evidence, alerts, incidents and reports are associated with\nit by ``engagement_id``. CLI, TUI, API and Web all reference this one model\n(ROADMAP ``WEB-B``) rather than a private per-interface notion of scope.",
        "properties": {
          "authorization_reference": {
            "default": "",
            "title": "Authorization Reference",
            "type": "string"
          },
          "client": {
            "default": "",
            "title": "Client",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "engagement_id": {
            "title": "Engagement Id",
            "type": "string"
          },
          "metadata": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Metadata",
            "type": "object"
          },
          "name": {
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "schema_name": {
            "const": "olympus.engagement",
            "default": "olympus.engagement",
            "title": "Schema Name",
            "type": "string"
          },
          "schema_version": {
            "const": "1.0.0",
            "default": "1.0.0",
            "title": "Schema Version",
            "type": "string"
          },
          "scope": {
            "$ref": "#/components/schemas/EngagementScope"
          },
          "status": {
            "$ref": "#/components/schemas/EngagementStatus",
            "default": "active"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "title": "Tags",
            "type": "array"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "name",
          "scope"
        ],
        "title": "Engagement",
        "type": "object"
      },
      "EngagementList": {
        "additionalProperties": false,
        "description": "The engagements visible over the API, newest first.",
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "engagements": {
            "items": {
              "$ref": "#/components/schemas/Engagement"
            },
            "title": "Engagements",
            "type": "array"
          },
          "schema_name": {
            "const": "olympus.themis-engagement-list",
            "default": "olympus.themis-engagement-list",
            "title": "Schema Name",
            "type": "string"
          },
          "schema_version": {
            "const": "1.0.0",
            "default": "1.0.0",
            "title": "Schema Version",
            "type": "string"
          }
        },
        "required": [
          "count",
          "engagements"
        ],
        "title": "EngagementList",
        "type": "object"
      },
...
      "EngagementStatus": {
        "description": "Lifecycle state of an engagement (the top-level assessment container).",
        "enum": [
          "planned",
          "active",
          "closed"
        ],
        "title": "EngagementStatus",
        "type": "string"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "JobList": {
        "additionalProperties": false,
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "jobs": {
            "items": {
              "$ref": "#/components/schemas/ThemisJob"
            },
            "title": "Jobs",
            "type": "array"
          },
          "schema_name": {
            "const": "olympus.themis-job-list",
            "default": "olympus.themis-job-list",
            "title": "Schema Name",
            "type": "string"
          },
          "schema_version": {
            "const": "2.0.0",
            "default": "2.0.0",
            "title": "Schema Version",
            "type": "string"
          }
        },
        "required": [
          "count",
          "jobs"
        ],
        "title": "JobList",
        "type": "object"
      },
      "JobState": {
        "description": "Lifecycle state of one durable job.\n\nThe terminal states are deliberately distinct: collapsing them into\n``failed`` hides whether a scan broke, was refused, ran out of time, or\nnever ran at all.",
        "enum": [
          "queued",
          "running",
          "succeeded",
          "partial",
          "failed",
          "timed_out",
          "cancelled",
          "policy_denied"
        ],
        "title": "JobState",
        "type": "string"
      },
      "JobSubmission": {
        "additionalProperties": false,
        "description": "Bounded request body for one authorized job.",
        "properties": {
          "authorized": {
            "title": "Authorized",
            "type": "boolean"
          },
          "engagement_id": {
            "anyOf": [
              {
                "pattern": "^ENG-\\d{4}-\\d{5}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Engagement Id"
          },
          "idempotency_key": {
            "anyOf": [
              {
                "pattern": "^[A-Za-z0-9][A-Za-z0-9_.:-]{0,127}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Idempotency Key"
          },
          "max_attempts": {
            "default": 1,
            "maximum": 10.0,
            "minimum": 1.0,
            "title": "Max Attempts",
            "type": "integer"
          },
          "scanner": {
            "pattern": "^[a-z0-9][a-z0-9-]{0,63}$",
            "title": "Scanner",
            "type": "string"
          },
          "scope_id": {
            "pattern": "^[A-Za-z0-9][A-Za-z0-9_.-]{0,127}$",
            "title": "Scope Id",
            "type": "string"
          },
          "target": {
            "maxLength": 2048,
            "minLength": 1,
            "title": "Target",
            "type": "string"
          },
          "target_kind": {
            "default": "host",
            "enum": [
              "host",
              "domain",
              "url"
            ],
            "title": "Target Kind",
            "type": "string"
          }
        },
        "required": [
          "scanner",
          "target",
          "scope_id",
          "authorized"
        ],
        "title": "JobSubmission",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "in": "header",
        "name": "X-Olympus-API-Key",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Authorized specialist-engine orchestration and evidence lifecycle.",
    "title": "Olympus THEMIS API",
    "version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/capabilities": {
      "get": {
        "operationId": "capabilities_api_v1_capabilities_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Capabilities Api V1 Capabilities Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Capabilities",
        "tags": [
          "capabilities"
        ]
      }
    },
    "/api/v1/engagements": {
      "get": {
        "operationId": "list_engagements_api_v1_engagements_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EngagementList"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "List Engagements",
        "tags": [
          "engagements"
        ]
      }
    },
    "/api/v1/engagements/{engagement_id}": {
      "get": {
        "operationId": "get_engagement_api_v1_engagements__engagement_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "engagement_id",
            "required": true,
            "schema": {
              "title": "Engagement Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Engagement"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Get Engagement",
        "tags": [
          "engagements"
        ]
      }
    },
    "/api/v1/jobs": {
      "get": {
        "operationId": "list_jobs_api_v1_jobs_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 1000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "state",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/JobState"
                },
                {
                  "type": "null"
                }
              ],
              "title": "State"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobList"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "List Jobs",
        "tags": [
          "jobs"
        ]
      },
      "post": {
        "operationId": "submit_api_v1_jobs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobSubmission"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ThemisJob"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Submit",
        "tags": [
          "jobs"
        ]
      }
    },
    "/api/v1/jobs/{job_id}": {
      "get": {
        "operationId": "get_job_api_v1_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ThemisJob"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Get Job",
        "tags": [
          "jobs"
        ]
      }
    },
    "/api/v1/jobs/{job_id}/cancel": {
      "post": {
        "operationId": "cancel_job_api_v1_jobs__job_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ThemisJob"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Cancel Job",
        "tags": [
          "jobs"
        ]
      }
    },
...
  }
}